website basic functionality

+++ RFID website/index.js
@@ -1,25 +1,20 @@
-
-//THIS IS WHAT IS WORKING
-/*
-How it works:
-this javascript code uses "promises", which are objects that are basically incomplete, waiting to finish exicuting. 
-That is why "await" is used in front of all the "navigator.xxxx" functions. This allows many of these to run 
-at the same time without blocking.
-
-*can send and receive data when using multiple test function. but when using independent functions it breaks after first send. then disconnects.
-
-
-
-*/
+/**
+ * RFID Cloner project
+ * 
+ * Website controls and stores data from the cloner
+ * 
+ * 
+ */
 
 //GLOBAL VARIABLES: simplifies using functions with HTML
 let cloner;
 let server;
 let service;
-//UUID's for characteristics for cloner, one for transmit data, one for receive, one for command to scan
-let write_badge_to_cloner_ID = "0000180f-0000-1000-8000-000000000001";
-let receive_badge_from_cloner_ID = "0000180f-0000-1000-8000-000000000002";
-let scan_badge_command = "0000180f-0000-1000-8000-000000000003";
+
+let write_badge_to_cloner_ID = 0x0001;
+let receive_badge_from_cloner_ID = 0x0002;
+let scan_badge_command = 0x0003;
+let turn_off_device_command = 0x0004;
 //encoder and decoder for converting byte data into strings
 decoder = new TextDecoder();
 encoder = new TextEncoder();
@@ -27,105 +22,161 @@ encoder = new TextEncoder();
 let cloner_transmit_characteristic;
 let cloner_receive_characteristic;
 let cloner_scan_command_characteristic;
+let cloner_turn_off_characteristic;
 //variables for sending and receiving data
 let cloner_to_web_data;
 let web_to_cloner_data;
 
 
 
-
-
-//FUNCTIONS
+//FUNCTIONS***********************************************************************************
 
 
 
 //gets cloner connected to website and gets all characteristics
-async function connect_to_cloner(){
+async function connect_to_cloner() {
     //device find
     cloner = await navigator.bluetooth.requestDevice({
-    acceptAllDevices: true,
-    optionalServices: ['battery_service']
-  });
+        acceptAllDevices: true,
+        optionalServices: ['battery_service']
+    });
 
-//connect to cloner peripheral
-console.log('Connecting to Cloner Peripheral...');
-server = await cloner.gatt.connect();
-//get cloner service
-console.log('Getting main service...');
-service = await server.getPrimaryService('battery_service');
+    //connect to cloner peripheral
+    console.log('Connecting to Cloner Peripheral...');
+    server = await cloner.gatt.connect();
+    //get cloner service
+    console.log('Getting main service...');
+    //service = await server.getPrimaryService('battery_service');
+    service = await server.getPrimaryService('battery_service');
 
-//get characteristics from cloner. these characteristics will link to buttons in website. they will each serve as different command for device
-console.log('Getting characteristics');
-cloner_transmit_characteristic = await service.getCharacteristic(receive_badge_from_cloner_ID);
-cloner_receive_characteristic = await service.getCharacteristic(write_badge_to_cloner_ID);
-cloner_scan_command_characteristic = await service.getCharacteristic(scan_badge_command);
-
-console.log("DONE!");
+    //get characteristics from cloner. these characteristics will link to buttons in website. they will each serve as different command for device
+    console.log('Getting characteristics');
+    cloner_transmit_characteristic = await service.getCharacteristic(receive_badge_from_cloner_ID);
+    cloner_receive_characteristic = await service.getCharacteristic(write_badge_to_cloner_ID);
+    cloner_scan_command_characteristic = await service.getCharacteristic(scan_badge_command);
+    cloner_turn_off_characteristic = await service.getCharacteristic(turn_off_device_command);
+    //publish device name to website
+    document.getElementById("bluetooth_device_name").innerHTML = cloner.name;
+    //start notification services for cloner
+    setup_RFID_notifications();
+    console.log("DONE!");
 }
 
 
-//send badge number to cloner
-async function send_data_to_cloner(){
-    try{
-    //let new_RFID_UID = document.getElementById("text_box").value;
-    let val = document.getElementById("text_box").value;
-    //convert to byte array
-    byte_buff = await encoder.encode(val);
-    await cloner_receive_characteristic.writeValue(byte_buff);
+
+
+//send badge number to cloner via button push
+async function send_data_to_cloner() {
+    try {
+        //let new_RFID_UID = document.getElementById("text_box").value;
+        let val = document.getElementById("text_box").value;
+        //save custom UID to list of ID's
+        storeRFIDCode(val);
+        
+        //convert to byte array
+        byte_buff = await encoder.encode(val);
+        await cloner_receive_characteristic.writeValue(byte_buff);
+        console.log(byte_buff);//test
     }
-    catch(error){}
-    
-    
+    catch (error) {
+        console.log(error.message);
+    }
 
 }
 
-//get badge from cloner
-async function receive_data_from_cloner(){
+
+
+
+//read cloners characteristic via button push
+async function receive_data_from_cloner() {
+
     //read cloner characteristic
     cloner_to_web_data = await cloner_transmit_characteristic.readValue();
-    console.log(decoder.decode(cloner_to_web_data));
+
+    let val = decoder.decode(cloner_to_web_data);
+    console.log(val);
+    //store new badge to local storage
+    storeRFIDCode(val);
     
+
+    console.log(localStorage.getItem('rfidCodes'));//test to see if storage works
+    //localStorage.removeItem('rfidCodes');//test only!
+
 }
+
+
+
+
+//start RFID cloner notification: this will start an event handler to autimaticaly listen for cloner data. automatically saves new data to JSON file
+async function setup_RFID_notifications() {
+
+    if (cloner_transmit_characteristic.properties.notify) {
+
+        //get updated characteristic if notification received
+        try {
+            await cloner_transmit_characteristic.startNotifications();
+        }
+        catch (error) { }
+
+        //add event listener to characteristic, then if notification change val
+        cloner_transmit_characteristic.addEventListener(
+            "characteristicvaluechanged",
+
+            //aero function event listener activates when notification arrives
+            async (event_handler) => {
+
+                try {
+                    val = event_handler.target.value;
+                    val = decoder.decode(val);
+                    document.getElementById("RFID_Badge_number").innerHTML = val;
+                    //store UID to JSON data
+                    storeRFIDCode(val);
+                    
+                }
+                catch (DOMException) { }
+
+                console.log(localStorage.getItem('rfidCodes'));//TEST
+                console.log(val);
+
+            }
+        )
+    }
+}
+
+
+
 
 //commands cloner to scan badge
-function cloner_command_scan(){
+async function cloner_command_scan() {
 
+    try {
+        //Cloner waits for data to arrive on this characteristic. Doesnt matter what the data is, just that it arrives.
+        let data = await encoder.encode("writetoRFID_");
+        await cloner_scan_command_characteristic.writeValue(data);
+    }
+    catch (error) { }
 }
 
 
 
-
+//function to write to the on off characteristic. turns off device 
+async function cloner_turn_off(){
+    let data =  await encoder.encode("turnoffnow_");
+    await cloner_turn_off_characteristic.writeValue(data);
+}
 
 
 
 
 /**
-* Function to store an RFID code in local storage.
-* Attempts to handle errors.
-* @param {string} code - The RFID code to store.
-*/
-function storeRFIDCode(code) {
-let existingCodes;
-
-try {
-
-    existingCodes = localStorage.getItem('rfidCodes')
-    existingCodes = existingCodes ? JSON.parse(existingCodes) : [];
-
-} catch (error) {
-
-    existingCodes = []
-}
-
-existingCodes.push(code);
-
-try {
-
-    localStorage.setItem('rfidCodes', JSON.stringify(existingCodes));
-} catch (error) {
-
-    console.error('Failed to save RFID code: ', error);
-}
+ * Function to clear JSON data from web browser
+ * takes no parameters
+ */
+function clear_saved_data() {
+    localStorage.removeItem("rfidCodes");
+    console.log("cleared");
+    //clear drop down list 
+    update_badge_list()
 }
 
 
@@ -139,25 +190,38 @@ try {
 function storeRFIDCode(code) {
     let existingCodes;
 
+    //retreive badge numbers from storage and put into array
     try {
-
         existingCodes = localStorage.getItem('rfidCodes')
         existingCodes = existingCodes ? JSON.parse(existingCodes) : [];
-
-    } catch (error) {
-
+    }
+    //if no list available, make new array
+    catch (error) {
         existingCodes = []
     }
 
-    existingCodes.push(code);
+    //if new code not in list, add and save
+    if (!existingCodes.includes(code)) {
+        //add new code to the array of codes
+        existingCodes.push(code);
 
-    try {
+        //save codes
+        try {
+            localStorage.setItem('rfidCodes', JSON.stringify(existingCodes));
+        } catch (error) {
 
-        localStorage.setItem('rfidCodes', JSON.stringify(existingCodes));
-    } catch (error) {
-
-        console.error('Failed to save RFID code: ', error);
+            console.error('Failed to save RFID code: ', error);
+        }
+        //update drop down list
+        update_badge_list();
     }
+    //badge in list already, ignore
+    else {
+        let message = "UID already in list!";
+        console.log(message);
+    }
+
+
 }
 
 /**
@@ -185,91 +249,52 @@ function retrieveAllCodes() {
 
 
 
+//drop down list functions: 
+//the drop down list takes all its info from local storage to keep them in sync. 
 
-
-
-
-//**************Old functions not in use ************************/
-
-
-
-async function test_multiple_characteristics(){
-
-    //device find
-     cloner = await navigator.bluetooth.requestDevice({
-        acceptAllDevices: true,
-        optionalServices: ['battery_service']
-      });
+function fill_badge_list(badges,select_list){
     
-    
-    console.log('Connecting to Cloner Peripheral...');
-    server = await cloner.gatt.connect();
-    
-    console.log('Getting main service...');
-    service = await server.getPrimaryService('battery_service');
-    
-    console.log('Getting Battery Level Characteristic...');
-    //const characteristics = await service.getCharacteristics(); multiple characteristics (not using at moment)
-    //get characteristics from cloner. these characteristics will link to buttons in website. they will each serve as different command for device
-    cloner_transmit_characteristic = await service.getCharacteristic(receive_badge_from_cloner_ID);
-    cloner_receive_characteristic = await service.getCharacteristic(write_badge_to_cloner_ID);
-    cloner_scan_command_characteristic = await service.getCharacteristic(scan_badge_command);
-    
-    // let rx_data = (await cloner_receive_characteristic.readValue()).getUint8();
-    // let scan_now = (await cloner_scan_command_characteristic.readValue()).getUint8();
-    //while(true){
-    cloner_to_web_data = await cloner_transmit_characteristic.readValue();
-    console.log(decoder.decode(cloner_to_web_data));
-    //}
-    
-    while(true){
-        let new_RFID_UID = "so we do it";
-        await cloner_receive_characteristic.writeValue(encoder.encode(new_RFID_UID)); 
+    //fill select list with new badges
+    for(let i = 0;i<badges.length; i++){
+        //new option for list (empty)
+        let new_option = document.createElement("option");
+        //fill option with data
+        new_option.textContent = badges[i];
+        new_option.value = badges[i];
+        //add option to select list in HTML
+        select_list.appendChild(new_option);
     }
-    
+}
+
+//helper function to clear badge list
+function clear_badge_list(badge_list){
+    len = badge_list.options.length-1;
+    //loop through drop down list and clear it
+    for(let i = len; i>=0;i--){
+        badge_list.remove(i);
     }
+}
 
-    
+//driver function to fill badge list
+function update_badge_list(){
+    //read badges from memory
+    let badges = retrieveAllCodes();
+    //dropdown menu from HTML
+    let badge_list = document.getElementById("badges_list");
+    //clear data
+    clear_badge_list(badge_list);
+    //refill with updated data
+    fill_badge_list(badges, badge_list);
+    console.log(badges);//test
+}
 
-async function test(){
-    //device find
-    let device = await navigator.bluetooth.requestDevice({
-        acceptAllDevices: true,
-        optionalServices: ['battery_service']
-      });
-    document.getElementById("bluetooth_device_name").innerHTML = device.name;
 
-    console.log('Connecting to GATT Server...');
-    const server = await device.gatt.connect();
-    
-    console.log('Getting Battery Service...');
-    const service = await server.getPrimaryService('battery_service');
-    
-    console.log('Getting Battery Level Characteristic...');
-    const characteristic = await service.getCharacteristic('battery_level');
-   
-    //if characteristic has notifications enabled
-    if(characteristic.properties.notify){
-        //add event listener to characteristic, then if notification change val
-        characteristic.addEventListener(
-            "characteristicvaluechanged", 
-            //aero function: event is an event listener. this could be replaced with "_" or blank. It only serves as placeholder
-            async (event_handler) => {
-                //print new val to web page
-                val = await characteristic.readValue();
-                document.getElementById("RFID_Badge_number").innerHTML = val.getUint8(0);
-            }
-        )
-        //get updated characteristic if notification received
-        await characteristic.startNotifications();
-        
-    }
-    //you should have notifications enabled for live data!
-    else {
-        window.alert("your bluetooth device doesnt have notifications enabled!");
-    }
-    
-    console.log("done");
-
-    
+//function to fill text box with selection from drop down list
+function selection_to_text_box(){
+    //HTML text box and drop down list
+    let select_box = document.getElementById("badges_list");
+    let text_box = document.getElementById("text_box");
+    //take drop down list selection value and send to text box
+    let selection = select_box.value;
+    text_box.value = selection;
 }
